bg_flat/requirements.txt
# FITS文件背景抽取工具依赖

# 核心天文数据处理
astropy>=5.0.0

# 背景估计和光度测量
photutils>=1.5.0

# 数值计算
numpy>=1.20.0

# 图像处理和可视化
matplotlib>=3.5.0
Pillow>=8.0.0

# 统计和科学计算
scipy>=1.7.0

# 可选：更高级的图像处理
# scikit-image>=0.19.0
# opencv-python>=4.5.0
